refactor: 분석 컴포넌트 병합 #24
